chore: separate 'changeset' crate integration tests
