+++ crates/gitbutler-changeset/tests/changeset.rs
@@ -1,0 +1,3 @@
+mod diff;
+mod signature;
+mod span;
